Optimize types dir
